// src/core/enums/football-teams.enum.ts
export enum FOOTBALL_TEAMS {
  // Premier League
  MAN_UNITED = 'Manchester United',
  MAN_CITY = 'Manchester City',
  LIVERPOOL = 'Liverpool',
  CHELSEA = 'Chelsea',
  ARSENAL = 'Arsenal',
  TOTTENHAM = 'Tottenham Hotspur',
  WEST_HAM = 'West Ham United',
  EVERTON = 'Everton',
  ASTON_VILLA = 'Aston Villa',
  WOLVERHAMPTON = 'Wolverhampton Wanderers',
  // La Liga
  REAL_MADRID = 'Real Madrid',
  BARCELONA = 'Barcelona',
  ATLETICO_MADRID = 'Atletico Madrid',
  // Bundesliga
  BAYERN_MUNICH = 'Bayern Munich',
  BORUSSIA_DORTMUND = 'Borussia Dortmund',
  // Serie A
  JUVENTUS = 'Juventus',
  AC_MILAN = 'AC Milan',
  INTER_MILAN = 'Inter Milan',
  NAPOLI = 'Napoli',
  // National teams
  VIETNAM = 'Vietnam',
  THAILAND = 'Thailand',
  CAMBODIA = 'Cambodia',
  LAOS = 'Laos',
  PHILIPPINES = 'Philippines',
  MALAYSIA = 'Malaysia',
  INDONESIA = 'Indonesia',
  SINGAPORE = 'Singapore',
  MYANMAR = 'Myanmar',
  BRUNEI = 'Brunei',
}